--- doppler.yaml
@@ -1,3 +1,0 @@
-setup:
-  project: nova
-  config: dev
